refactor(main): toast helper を diff_viewer module に分離
auto-dismiss と toast push の副作用を src/app/diff_viewer/toast.rs に移し、main.rs を callback 配線へ寄せる。

Refs: #224
